% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapply.R
\name{cmapply}
\alias{cmapply}
\title{Apply a function between pairs of vectors}
\usage{
cmapply(FUN = function(x, y) sum(x \%in\% y), show = "all", ...)
}
\arguments{
\item{FUN}{the function to be applied between all possible pairs of vectors passed to ...}

\item{show}{Whether to display upper or lower triangle of values--either 'all', 'upper', or 'lower'}

\item{...}{named or unnamed vectors}
}
\value{
a matrix
}
\description{
Apply a function between pairs of vectors
}
\examples{
samples <- replicate(3, sample(LETTERS, 20, replace = TRUE))
sui <- function(x, y) sum(unique(x) \%in\% unique(y))
sm <- cmapply(s1 = samples[,1], s2 = samples[,2], s3 = samples[,3], FUN = sui)
}
