% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_heatmap.R
\name{combo_heatmap}
\alias{combo_heatmap}
\title{Plot heatmap result of pmapply}
\usage{
combo_heatmap(combo_mat, gradient_low = "lightgrey",
  gradient_high = "red")
}
\arguments{
\item{combo_mat}{a matrix generated from `pmapply`}

\item{gradient_low}{color for low values}

\item{gradient_high}{color for high values}
}
\description{
Plot heatmap result of pmapply
}
\examples{
\dontrun{
cormat <- pmapply(iris[,c(1:4)], function(x, y) cor(x, y))
combo_heatmap(cormat)
}
}
