% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minmax_combos.R
\name{get_minmax_combos}
\alias{get_minmax_combos}
\title{Get the minimum and maximum combinations from pmapply matrix}
\usage{
get_minmax_combos(combo_mat)
}
\arguments{
\item{combo_mat}{a matrix generated from `pmapply`}
}
\value{
a list object of class 'minmaxCombo'
}
\description{
Get the minimum and maximum combinations from pmapply matrix
}
\examples{
\dontrun{
cormat <- pmapply(iris[,c(1:4)], function(x, y) cor(x, y))
get_minmax_combos(cormat)
}
}
