% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmapply.R
\name{pmapply}
\alias{pmapply}
\title{Apply a function between pairs of vectors}
\usage{
pmapply(x, FUN = function(x, y) sum(x \%in\% y), show = "all")
}
\arguments{
\item{x}{a list or data frame}

\item{FUN}{the function to be applied between all possible pairs of vectors passed to ...}

\item{show}{Whether to display upper or lower triangle of values--either 'all', 'upper', or 'lower'}
}
\value{
a matrix
}
\description{
Apply a function between pairs of vectors
}
\examples{
samples <- replicate(3, sample(LETTERS, 20, replace = TRUE))
l <- list(samples[,1], samples[,2], samples[,3])
pmapply(l)
pmapply(iris[,c(1:4)], function(x, y) cor(x, y))
pmapply(iris[,c(1:4)], function(x, y) sum(x < y))
}
