% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minmax_combos.R
\name{print}
\alias{print}
\alias{print.minmaxCombo}
\title{Print the output of get_minmax_combos}
\usage{
\method{print}{minmaxCombo}(combo_mat)
}
\arguments{
\item{combo_mat}{a matrix generated from `pmapply`}
}
\description{
Print the output of get_minmax_combos
}
\section{Methods (by class)}{
\itemize{
\item \code{minmaxCombo}: Prints output of get_minmax_combos
}}

